requirements.txt
fastapi==0.115.8
psycopg2-binary  
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic_core==2.10.1
PyJWT==2.10.1
python-dateutil==2.8.2
python-dotenv
python-multipart==0.0.20
redis==5.0.0
requests==2.28.1
urllib3==1.26.13
uvicorn==0.34.0
watchdog==2.1.9
sqlalchemy
passlib
